Use a more idiomatic loop idiom
I use the term "idiomatic" here with caution. I don't think this idiom is actually all that widespread.
But maybe it should be.

Note that we are actually "lying" to the type system at this point. The `!` is asserting that the shift will
never be undefined, which is what makes the assignability to `elem` work. But at some point, the
`shift` will give back an `undefined`, which will halt the loop. So `elem` will (unsoundly) contain an
`undefined`, but after seeing it, we never use it again. So it's "fine".